+++ telegrambot/bot_views/__init__.py
@@ -1,0 +1,1 @@
+from telegrambot.bot_views.login import LoginBotView  # NOQA
