+++ telegrambot/bot_views/generic/__init__.py
@@ -1,0 +1,5 @@
+from telegrambot.bot_views.generic.base import TemplateCommandView  # noqa
+from telegrambot.bot_views.generic.compound import ListDetailCommandView  # noqa
+from telegrambot.bot_views.generic.detail import DetailCommandView  # noqa
+from telegrambot.bot_views.generic.list import ListCommandView  # noqa
+from telegrambot.bot_views.generic.responses import (TextResponse, KeyboardResponse)  # noqa
