+++ telegrambot/bot_views/generic/detail.py
@@ -1,4 +1,4 @@
-from telegrambot.generic.base import TemplateCommandView
+from telegrambot.bot_views.generic.base import TemplateCommandView
 from django.core.exceptions import ImproperlyConfigured, ObjectDoesNotExist,\
     FieldError
