--- telegrambot/generic/__init__.py
@@ -1,5 +1,0 @@
-from telegrambot.generic.base import TemplateCommandView  # noqa
-from telegrambot.generic.compound import ListDetailCommandView  # noqa
-from telegrambot.generic.detail import DetailCommandView  # noqa
-from telegrambot.generic.list import ListCommandView  # noqa
-from telegrambot.generic.responses import (TextResponse, KeyboardResponse)  # noqa
