+++ telegrambot/handlers/__init__.py
@@ -1,0 +1,4 @@
+from telegrambot.handlers.command import command  # noqa
+from telegrambot.handlers.message import message  # noqa
+from telegrambot.handlers.regex import regex  # noqa
+from telegrambot.handlers.unknown_command import unknown_command  # noqa
