+++ telegrambot/test/factories/__init__.py
@@ -1,0 +1,5 @@
+from telegrambot.test.factories.auth import AuthTokenFactory  # noqa
+from telegrambot.test.factories.bot import BotFactory  # noqa
+from telegrambot.test.factories.telegram_lib import (UserLibFactory, ChatLibFactory,  # noqa
+                                                     MessageLibFactory, UpdateLibFactory)  # noqa
+from telegrambot.test.factories.user import UserWebFactory  # noqa
